Include object ref size while constructing children
Previously we assumed a ref_size of 1 while traversing children. This
fails around the 4KB file size where the ref size jumps to 2. Now that
we allow ref sizes to be arbitrarily (`usize::MAX`) large, it should be
able to be handle pretty much any file :)

Tried with a 66KB file (ref_size 4)